feat: generalize `DenseFilterExpr` (#8)
# Rationale for this change
It is necessary to move provable projections into the prover so that
provable queries do not need postprocessing. This is the first PR of the
change which is to move provable projections to the prover for filter
queries.
<!--
Why are you proposing this change? If this is already explained clearly
in the linked Jira ticket then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?
- Move `Expression` -> `ProvableExprPlan` to a separate struct,
`ProvableExprPlanBuilder` since the process is no longer for where
clauses only.
- Add `EnrichedExpr` to be equivalent to the provable version of
[Expr](https://docs.rs/datafusion/latest/datafusion/logical_expr/enum.Expr.html#method.alias)
in DataFusion for tamperproof queries.
- Allow arbitrary aliased provable expressions as result expressions in
dense filters.
- Remove constraints in `QueryContext` that prevents constants from
being used in select.
<!--
There is no need to duplicate the description in the ticket here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

# Are these changes tested?
Existing tests should pass. New tests will be added to cover nontrivial
results such as `select a and b as c, d, 4 as f from tab where d = e`.
<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->