perf: Dory's compute_T_vec_prime should be able to use Blitzar's fixed MSM (PROOF-876) (#7)
# Rationale for this change
When creating a Dory proof the `compute_T_vec_prime` function requires
multiple MSMs. The current implementation of this function only supports
the use of the Arkworks library, which only runs on the CPU. In an
effort to improve the performance of Dory, this PR introduces the
Blitzar fixed MSM to `compute_T_vec_prime` when appropriate to perform
the MSM computations.

The commit focusing only on performance changes is available
[here](https://github.com/spaceandtimelabs/sxt-proof-of-sql/commit/ba506b0b5c4be809ca424e07423c67f15fee6c15).

# What changes are included in this PR?
- `compute_T_vec_prime` now supports Blitzar's fixed MSM if the feature
flag `"blitzar"` is turned on. The CPU version using Arkworks is still
supported.
- The `transpose_column` and related functions appearing in the
`dory_commitment_helper_gpu` class are moved into a separate class,
`transpose`.
- Added a reference to the `blitzar_handle` to the `ProverSetup`.
- Removed `par_iter()` in `eval_vmv_re.rs` because `*mut
blitzar_sys::sxt_multiexp_handle` cannot be shared between threads
safely.

# Are these changes tested?
Yes, on both CPU and GPU.