feat: add aggregation-related intermediate AST conversions (#32)
# Rationale for this change
This is the first PR in generalization of `GroupByExpr`. Following
conventions from
[datafusion](https://docs.rs/datafusion/latest/datafusion/logical_expr/enum.Expr.html)
we allow `AggregateExpr` so that provable aggregations can be expressed
in `ProvableExprPlan` and provide relevant conversion.
<!--
Why are you proposing this change? If this is already explained clearly
in the linked Jira ticket then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?
- add `AggregateExpr` which allows expressions such as `sum(a+b)` to
become `ProvableExprPlan`
- allow `Expression` -> `ProvableExprPlan` for aggregations
<!--
There is no need to duplicate the description in the ticket here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

# Are these changes tested?
Existing tests should pass.
<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->