feat: add `ProjectionExpr` (#41)
# Rationale for this change
We need to add `ProjectionExpr` both as a faster and simpler alternative
to `DenseFilterExpr` when there is no `WHERE` clause and to provide one
mechanism to further generalize `GROUP BY` i.e. adding a
`ProjectionExpr` after `GroupByExpr` without changing `GroupByExpr`
itself.
<!--
Why are you proposing this change? If this is already explained clearly
in the linked Jira ticket then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?
Add `ProjectionExpr`.
<!--
There is no need to duplicate the description in the ticket here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

# Are these changes tested?
Yes.
<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->