fix: fix postprocessing projection in provable `GROUP BY`s & allow `WHERE` clause to be omitted (#50)
# Rationale for this change
Recently we found weird postprocessing bugs for group bys. What we found
is that
1. Currently provable `GROUP BY`s without a `WHERE` clause are not
supported even though to support it is trivial.
2. Currently if a `GROUP BY` clause is provable the `ResultExpr` still
incorrectly contains aggregate functions.
<!--
Why are you proposing this change? If this is already explained clearly
in the linked Jira ticket then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?
- fix postprocessing projection with provable `GROUP BY`
- allow provable group by without `WHERE` clause
- enable 3 ignored tests
<!--
There is no need to duplicate the description in the ticket here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

# Are these changes tested?
Yes
<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
3. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->