feat: time unit try from string (#45)
# Rationale for this change

In DDL, a Timestamp column can be instantiated with the following
syntax:

```sql
TIMESTAMP(6)
```

where 6 represents millisecond precision. It appears that the gateway by
default supports only nanosecond precision on DDL, however if this
changes in the future, we will need a way to parse the time unit from a
DDL message in proofs-service. This PR contains a simple string parser
that allows for selection of a supported precision for timestamp
columns.

# What changes are included in this PR?

- [x] Implements TryFrom &str for TimeUnit
- [x] updates Display implementation for TImestamp column type to fix a
backwards print that is occuring:

```
column type TIMESTAMP(TIMEUNIT: UTC, TIMEZONE: NANOSECONDS (PRECISION: 9)) 
``` 

# Are these changes tested?

yes