fix: fix decimal comparison and aggregation (#51)
# Rationale for this change
It is necessary for us to be able to do comparison and aggregation on
decimal columns.
<!--
Why are you proposing this change? If this is already explained clearly
in the linked Jira ticket then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?
- add `signed_cmp` to `Scalar` definition to do signed Scalar comparison
(i.e. `S::MAX_SIGNED` is the largest number and everything "beyond" it
wraps around into negative numbers)
- allow decimal comparison
- allow `SUM` of decimals
<!--
There is no need to duplicate the description in the ticket here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

# Are these changes tested?
Yes.
<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->