build: upgrade to arrow 51 (#60)
# Rationale for this change

<!--
Why are you proposing this change? If this is already explained clearly
in the linked Jira ticket then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->
Our current version of arrow is quite old (version v45). Attempts to
update this in the past have been somewhat thwarted by dependent
projects depending on an older version of arrow and being stuck to it
due to some complicated dependency management. This is now being
resolved, and an arrow upgrade to proof-of-sql is required to resolve
it.

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the ticket here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->
- arrow dependency upgraded to 51
- deprecation warnings addressed

# Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->
These changes do not affect existing functionality which is verified by
existing tests.