refactor: add flexibility for dory verifier folding
In order to make the existing methods work with the dynamic dory commitment scheme, the method that folds the s tensors within the verify methods needs to be modified. In particular, it needs to have access to all the folding factors at a time in order to fold properly.

The folding of the s tensor is now done once within the `fold_scalars_0_verify` rather than accumulating within each round.
There are 2 main changes to accomplish this:
* During each round/call of `extended_dory_reduce_verify`, instead of mutating the s tensors, the `alpha` folding factors are added to the `state`.
* `fold_s_tensors_verify` is added as a parameter to `extended_dory_inner_product_verify` and `fold_scalars_0_verify`