feat: add naive `fold_dynamic_standard_basis_tensors`
The `fold_dynamic_standard_basis_tensors` is added and implemented using a naive mechanism. Tests are also added to ensure that any future optimizations are still valid.
This method is the verifier's equivalent of `compute_dynamic_standard_basis_vecs`. Ultimately, there is still wrapper code needed to use this method.