refactor: explicitly mention all enum variants in `ColumnType::precision_value` and `ColumnType::scale` (#76)
# Rationale for this change
This is a clean and explicit solution. Note that for all timestamp types
we have `scale = 0` since in #75 time unit is removed anyway. Also this
is a refactoring since it doesn't affect functionality.
<!--
Why are you proposing this change? If this is already explained clearly
in the linked Jira ticket then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?

- Explicitly mention all enum variants in `ColumnType::precision_value`
and `ColumnType::scale`

<!--
There is no need to duplicate the description in the ticket here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

# Are these changes tested?
Yes
<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->