feat: define `Add`, `Sub` and `Mul` for `OwnedColumn` (#74)
# Rationale for this change
This PR is split out from #70 since it has become too large. The actual
content of this PR (that is, what's not in #70) is here:
https://github.com/spaceandtimelabs/sxt-proof-of-sql/pull/74/commits/e1667622c3db3b45af75dd1f1bf167f3e2a646e5
<!--
Why are you proposing this change? If this is already explained clearly
in the linked Jira ticket then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?

- Define `Add`, `Sub` and `Mul` for `OwnedColumn`

<!--
There is no need to duplicate the description in the ticket here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

# Are these changes tested?
Yes.
<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->