feat: add `Into<BigInt>` to `Scalar` & improvement to `Scalar` in general (#78)
# Rationale for this change
This PR is a prerequisite of #77. 
<!--
Why are you proposing this change? If this is already explained clearly
in the linked Jira ticket then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?
- add implementation as a part of `scalar_conversion_to_int` macro which
is necessary since blanket implementation of `TryFrom` or `TryInto` is
not allowed for traits by Rust
- improve the macro so that `TryInto` is replaced with the more natural
`TryFrom`
- replace `std` with `core`
<!--
There is no need to duplicate the description in the ticket here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

# Are these changes tested?
Yes
<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->