ci: add CI and basic tests for Solidity (#120)
# Rationale for this change

In order to facilitate development of a Solidity verifier, we should be
running unit tests within the CI.

# What changes are included in this PR?

* `forge-script` is used to call solidity code from rust. This enables
writing tests where the test cases are generated by Rust. For example,
this is helpful when trying to test that a proof is verified. The proof
can be generated by rust, and verified in solidity. Because these are
included in the Rust test suite, nothing is added in CI to achieve this.
Some example tests are added.
* A `foundry.toml` project file is added to the top directory and `forge
test -vvv` is now in CI in order to run pure Solidity tests (functions
that start with `test`)
* `solhint` is now part of CI in order to enforce linting of Solidity
code.

# Are these changes tested?

Yes
* The example Rust+Solidity tests execute as expected.
* `forge test -vvv` fails on failed tests and shows up as failed in the
CI:
https://github.com/spaceandtimelabs/sxt-proof-of-sql/actions/runs/10598506903/job/29371258834
* `solhint` fails when a lint is cause and shows up as failed in the CI:
https://github.com/spaceandtimelabs/sxt-proof-of-sql/actions/runs/10598506903/job/29371261143