refactor!: separate expressions from execution plans && renames (#124)
# Rationale for this change
As the amount of expressions and execution plans increase we need to
seriously differentiate between the two. This PR is the first in a
series of PRs that will lead to `SliceExec` and `UnionExec`.
 
<!--
Why are you proposing this change? If this is already explained clearly
in the linked Jira ticket then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?
- rename `ProvableExpr` to `ProofExpr`
- rename `ProvableExprPlan` to `DynProofExpr`
- rename `AliasedProvableExprPlan` to `AliasedDynProofExpr`
- rename `ProvableExprPlanBuilder` to `DynProofExprBuilder`
- rename `ProofPlan` to `DynProofPlan`
- rename `ProofExecutionPlan` to `ProofPlan`
- move all `ProofExpr`s to `sql::proof_exprs`
- move all `ProofPlan`s to `sql::proof_plans`
- fix `Cargo.toml` warning in new Rust
 
<!--
There is no need to duplicate the description in the ticket here but it
is sometimes worth providing a summary of the individual changes in this
PR.
-->

# Are these changes tested?
Existing tests should pass
<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->