ci: install Nvidia drivers when running CI (PROOF-911) (#147)
# Rationale for this change
Blitzar now checks to the version of the Nvidia driver. It currently
throws an error if the version is under `nvidia-driver-560`, causing the
CI to fail. This PR installs the Nvidia driver before running tests with
the GPU.

# What changes are included in this PR?
- The Test Suite portion of the Github workflow now:
- Adds the NVIDIA package repository since the`nvidia-driver-560`
package is not yet available in the default Ubuntu repositories.
- Installs the`nvidia-driver-560` package when installing dependencies.

# Are these changes tested?
Yes, via Github CI.