feat: range check: decompose to words and get log derivative (#141)
# Rationale for this change

This PR brings initial support for the updated range check
functionality. It includes compartmentalized functions which handles the
decomposition of scalars into words, and it handles the addition of the
verifier challenge to each word and inverting the sum, forming the
logarithmic derivative of the word and challenge.

# What changes are included in this PR?

- [x] scalar decomposition
- [x] logarithmic derivative
- [x] update to commitable column to support u8

# Are these changes tested?

yes

- [x] we_can_decompose_small_scalars_to_words
- [x] we_can_decompose_large_scalars_to_words
- [x] we_can_obtain_a_logarithmic_derivative_from_a_small_scalar
- [x] we_can_obtain_a_logarithmic_derivative_from_a_large_scalar


# Split
- [ ] if alpha = P - word, then word + alpha mod P = 0 and has no
inverse. This breaks completeness and should be addressed