docs: add `missing_docs_in_private_items` lint (#156)
# Rationale for this change

In order to improve our internal documentation, we should be documenting
functionality. Enforcing this in CI simplifies review and development.

# What changes are included in this PR?

The `missing_docs_in_private_items` lint with the
`missing-docs-in-crate-items` option set to true. This option means that
missing docs are only warned if they have `pub(crate)` visibility. We
could set this to `false` to be even more stringent. However, this would
cause 473 more warnings for missing docs, and is likely a bit too much.

# Are these changes tested?

Yes