feat: refactor Commitment trait to return Vec<Self> instead of mutating slice (#184)
# Rationale for this change

Currently `Commitment::compute_commitments` mutates a slice buffer of commitments rather than returning a collection of the commitments. This is inherited from the behavior of the `blitzar` crate. However, it is more idiomatic for the implementations to return collections.

# What changes are included in this PR?

- Modified the `Commitment` trait's `compute_commitments` method to
return `Vec<Self>` instead of mutating a pre-allocated slice.
- Updated implementations for `NaiveCommitment`, `DoryCommitment`, and
`DynamicDoryCommitment` to align with the new trait definition.