refactor!: replace `compute_commitment_for_testing` with `RistrettoPoint::compute_commitments` (#186)
# Rationale for this change

The compute_commitment_for_testing function was a utility function to simplify computing commitments for testing purposes. Newer interfaces simplify this code, so this function is no longer needed, and adds clutter.

# What changes are included in this PR?

Replaced the `compute_commitment_for_testing` to
`RistrettoPoint::compute_commitments`,in the test case that was being
used

NOTE: this is a breaking change because `compute_commitment_for_testing` is removed, and it was public via the `test` feature.