feat: add helper methods to dynamic Dory structure (PROOF-919) (#198)
# Rationale for this change
The GPU implementation of the dynamic Dory commitment computation
requires a number of helper methods in the dynamic Dory structure. This
PR implements methods that will be used by the GPU implementation.

# What changes are included in this PR?
- Add `index_from_row_and_column` method which is the inverse of
`row_and_column_from_index` method. Given a `row` and `column` it will
provide the `index` of the original data. The method also handles cases
where `row` and `column` do not provide a valid `index`.
- Add `matrix_size` method which returns the width and height of the
dynamic Dory matrix structure that will hold the given data.
- Made helper methods `const` for compile-time evaluation.
- Add temporary `#[allow(dead_code)]` to address `cargo clippy` warnings
- these will be removed when the dynamic Dory commit computation is
implemented for the GPU.

# Are these changes tested?
Yes