fix: handle `IntermediateDecimal` precision and scale properly
`BigDecimal` gives "precision" and "scale" as 64-bit integers. `IntermediateDecimal` previously simply converted these to 8-bit integers, meaning that for precision or scale of 256 or 128 digits the code will return the wrong value. (Until recently, it actually wrapped the value, which is even more problematic.)
This commit fixes this by returning the proper value and requiring upstream usage to handle any casting explicitly as needed. To facilitate this, a conversion from `u64` to `Precision` is added.