feat: proversetup fileio (#295)
# Rationale for this change

This PR adds support for a conversion from pre-generated public
parameters and a blitzar handle into a ProverSetup. This helps skip an
expensive step to generate parameters and should significantly decrease
the amount of time to load the parameters. It serves as a step towards
file i/o for the parameters.

# What changes are included in this PR?

- [x] convert public params and blitzar handle into a proversetup

# Are these changes tested?

existing tests pass. a round trip test is forthcoming in a future PR.