fix: remove blitzar as a dev-dependency
Tests cannot currently be run on alternative architectures because they
enable the blitzar dependency.
However, this is not strictly necessary, because all tests using blitzar
are behind the blitzar feature flag.
So, this removes blitzar as a dev dependency.

However, doing so causes another issue.
It seems we are implicitly depending on the std feature of ark-serialize
via blitzar in tests.
So, running tests without blitzar also disables that feature.
This PR also updates the std feature to enable ark-serialize/std.