chore: add `OwnedColumn::try_coerce_scalar_to_numeric`
Additionally, the `ColumnCoercionError` type is added.

This method is used to convert a `Scalar` column to a numeric column.
This is useful because the Add/Sub/Mul nodes always return a `Scalar` column, rather than a column with the expected type.
In particular, this conversion is fallible because the columns could overflow the target type, or request an invalid conversion.