refactor!: use `OwnedTable` to serialize onto transcript
`postcard` does not port well to other language (e.g. Solidity). So, we need a simpler way to extend the transcript with the result.
Instead, we use `OwnedTable` to directly write to the transcript.

In order to do this, we replace `ProvableQueryResult` with `OwnedTable` everywhere. This requires updating the `exercise_verification` method.