refactor!: update scalar challenges to use mask (#474)
# Rationale for this change

The current mechanism for drawing a challenge is a bit opaque. This
simplifies it.

# What changes are included in this PR?

When drawing a challenge from the transcript, it is 32 bytes long.
Instead of using `[u64; 4]` to coerce it into a `Scalar`, it is more
straightforward to simply drop the first few bits. This can be
accomplished by masking the challenged by a `CHALLENGE_MASK` before
converting to a `Scalar`.

Additionally, this PR updates the `Scalar` implementations so that we
have a blanket implementation of `impl<T> Scalar for MontScalar<T>`.
This removes the need to manually figure out the constants for each
Scalar type.

# Are these changes tested?

Yes.