fix: put nova-snark dependency and hyperkzg module behind feature flag
proof-of-sql is currently not no_std compatible, as nova-snark is not
no_std compatible. nova-snark depends on halo2curves which depends on
std, as well as serde/std. So, to maintain the no_std compatibility of
proof-of-sql, the hyperkzg feature that depends on nova-snark needs to
be feature-gated for the time being.