refactor: extract logic for non-blitzar hyperkzg compute_commitments
Currently, only one of the hyperkzg commitment computation functions is
available at a time, the either blitzar version or the non-blitzar
version. As the non-blitzar version of the code is changing, it can be
useful to compare commitments produced by both versions and make sure
they are equal in tests. To accomplish this, this change takes all of
the logic out of the non-blitzar version of the code and makes this
inner function available for tests even when blitzar is enabled.