feat!: rearrange serialization order of TableCommitment fields
The current standard for serializing/deserializing table commitments
uses bincode, and the order of struct fields in bincode serialization
deterministically matches the order of the struct fields in rust. This
standard was chosen because, despite it being a binary format, custom
deserialization for it can easily be written in any language. So, at the
moment, the serialization order for table commitments is something
like..
1. commitments
2. column commitment metadata
3. range

Many use cases for custom deserialization don't care about the column
commitment metadata, but do care about the commitments and range. So, it
would be nice for these use cases to place the range first before the
dynamically sized items so that custom deserializers don't have to worry
about that dynamic sizing, and don't have to manually skip the
commitment metadata. This change reorders the serialization to be more
like...
1. range
2. commitments
3. column commitment metadata