feat: publicize QueryProof and VerifiableQueryResult fields
It is helpful for apis to expose proofs and query results separately.
These are both stored in a VerifiableQueryResult, which is the typical
result type for provers. However, not only are its fields private, but
the QueryProof type itself is private, so the two outputs cannot be
accessed individually pre-verification. This change simply publicizes
the necessary items to access the proof and query results individually.