refactor(sol): make `draw_challenges` return length-prefixed array
Previously, `draw_challenges` would allocate memory of length `count`,
fill that memory with `count` challenges,
and return a pointer to the first challenge.
This commit changes this so that it allocates an extra word,
starts the memory with a value of `count`,
and returns a pointer to that first value.

This aligns with the solidity memory layout for arrays.
While it is slightly less efficient, it is helpful for keeping the stack small.
Although, future optimizations may wish to revery this.