fix: use workspace for all dependencies in proof-of-sql-planner
Generally it's better to use workspace dependencies where possible to
avoid accidentally introducing duplicate dependencies as new crates are
developed. However, this is especially pertinent as we add automation to
release proof-of-sql-planner to crates.io. This automation failed
recently because the proof-of-sql dependency was defined with a path
instead of from the workspace.